requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary

djangorestframework-simplejwt
drf-yasg

pytest
pytest-django
pytest-cov
Pillow
gunicorn
django-widget-tweaks
django_prometheus
